Forward compatibility with PHPUnit 9
