+++ .travis.yml
@@ -16,8 +16,6 @@ matrix:
     - php: 7.3
     - php: 7.4
     - php: hhvm-3.18
-      install:
-        - composer require phpunit/phpunit:^5 --dev --no-interaction # requires legacy phpunit
   allow_failures:
     - php: hhvm-3.18
