+++ composer.json
@@ -32,7 +32,7 @@
         "clue/http-proxy-react": "^1.3",
         "clue/reactphp-ssh-proxy": "^1.0",
         "clue/socks-react": "^1.0",
-        "phpunit/phpunit": "^7.0 || ^6.4 || ^5.7 || ^4.8.35",
+        "phpunit/phpunit": "^9.0 || ^5.7 || ^4.8.35",
         "react/http": "^0.8"
     }
 }
